+++ src/App/Bundle/MainBundle/Services/Security/User/DoctrineWriter.php
@@ -4,6 +4,7 @@ namespace App\Bundle\MainBundle\Services\Security\User;
 
 use App\Bundle\MainBundle\Entity\Security\User as UserEntity;
 use App\Component\Security\User\Exception\CreationFailedException;
+use App\Component\Security\User\Exception\PromotionFailedException;
 use App\Component\Security\User\User;
 use App\Component\Security\User\WriterInterface;
 use Doctrine\DBAL\DBALException;
@@ -49,5 +50,36 @@ class DoctrineWriter implements WriterInterface
         } catch (DBALException $exception) {
             throw new CreationFailedException();
         }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function promote($identifier)
+    {
+        $entity = $this->getRepository()->findByUuid($identifier);
+
+        if ($entity === null || $entity->isModerator()) {
+            throw new PromotionFailedException();
+        }
+
+        $entity->setModerator(true);
+
+        try {
+            $this->entityManager->persist($entity);
+            $this->entityManager->flush();
+        } catch (DBALException $exception) {
+            throw new PromotionFailedException();
+        }
+    }
+
+    /**
+     * Retrieve the user entity repository
+     *
+     * @return \App\Bundle\MainBundle\Entity\Security\UserRepository
+     */
+    private function getRepository()
+    {
+        return $this->entityManager->getRepository('AppMainBundle:Security\User');
     }
 }