+++ src/App/Component/Security/User/WriterInterface.php
@@ -15,4 +15,13 @@ interface WriterInterface
      * @throws \App\Component\Security\User\Exception\CreationFailedException
      */
     public function create(User $user);
+
+    /**
+     * Promote the user with the given unique identifier as moderator
+     *
+     * @param string $identifier
+     *
+     * @throws \App\Component\Security\User\Exception\PromotionFailedException
+     */
+    public function promote($identifier);
 }