+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Paginate world records on the homepage and sort them by creation date
 - Account registration with a confirmation email
+- Command to promote a user as moderator
 
 ### Changed